feat: 전체 판매자 이력 및 상세 조회 페이지 구현
(cherry picked from commit 0dbbca9afbd1f87bdaad148e1a75366b5f1adbf5)
