+++ src/App.jsx
@@ -43,15 +43,9 @@ import SellerConcertManagementPage from './pages/seller/SellerConcertManagementP
 import AdminDashboard from './pages/admin/Dashboard';
 import AdminSellerManagement from './pages/admin/AdminSellerManagement';
 import SellerApproval from './pages/admin/SellerApproval.jsx';
+import ApplicationHistoryPage from './pages/admin/ApplicationHistoryPage.jsx';
 
-// --- 임시 관리자 페이지 컴포넌트들 (SellerApproval은 이제 실제 컴포넌트 사용) ---
-// const TempSellerApprovalPage = () => <div className="text-white p-4">판매자 승인 페이지 (임시)</div>;
-// const TempSellerManagementPage = () => (
-//     <div className="text-white p-4">판매자 관리 페이지 (임시)</div>
-// );
-const TempApplicationHistoryPage = () => (
-    <div className="text-white p-4">신청 이력 페이지 (임시)</div>
-);
+// --- 임시 관리자 페이지 컴포넌트 ---
 const TempSettingsPage = () => (
     <div className="text-white p-4">설정 페이지 (임시)</div>
 );
@@ -207,13 +201,16 @@ export default function App() {
                 >
                     <Route index element={<AdminDashboard />} />
                     <Route
+                        path="seller-management"
+                        element={<AdminSellerManagement />} // TempAdminSellerManagement -> AdminSellerManagement로 변경
+                    />
+                    <Route
                         path="seller-approvals"
                         element={<SellerApproval />} // TempSellerApprovalPage -> SellerApproval로 변경
                     />
-                    <Route path="sellers" element={<AdminSellerManagement />} />
                     <Route
                         path="history"
-                        element={<TempApplicationHistoryPage />}
+                        element={<ApplicationHistoryPage />} // TempApplicationHistoryPage -> ApplicationHistoryPage로 변경
                     />
                     <Route path="settings" element={<TempSettingsPage />} />
                 </Route>
